[#73109] Use Pages::SprintPlanning instead of Pages::Backlogs
